feat(step3): 매장 주문 데이터베이스 로직 작성

+++ src/test/java/kitchenpos/eatinorders/tobe/domain/order/vo/Quantity.java
@@ -3,9 +3,9 @@ package kitchenpos.eatinorders.tobe.domain.order.vo;
 import java.util.Objects;
 
 public class Quantity {
-    private final long value;
+    private final int value;
 
-    public Quantity(final long value) {
+    public Quantity(final int value) {
         this.value = value;
     }
